+++ addons/eventsheet/editor/dock/dock_ui_builder.gd
@@ -257,6 +257,7 @@ func build_ui() -> void:
 	_dock._viewport.span_edit_requested.connect(_dock._on_viewport_span_edit_requested)
 	_dock._viewport.navigate_requested.connect(_dock._navigate.navigate)
 	_dock._viewport.navigation_probe = _dock._navigate.can_navigate
+	_dock._viewport.rename_note_provider = _dock._variable_rename_note
 	_dock._viewport.editor_tool_action_requested.connect(_dock._editor_tool_bar.activate)
 	_dock._viewport.this_editor_action_requested.connect(_dock._this_editor_bar.activate)
 	_dock._viewport.ace_edit_requested.connect(_dock._on_viewport_ace_edit_requested)